ENH: Reverting to using long int instead of size_t

+++ kwsCheckBadCharacters.cxx
@@ -31,12 +31,12 @@ bool Parser::CheckBadCharacters(bool checkComments)
   
   unsigned long pos = 0;
   std::string::const_iterator it = buffer.begin();
-  size_t currentline = -1;
+  long int currentline = -1;
   while(it!= buffer.end())
     {
     if((*it)<0) 
       {
-      size_t linenum = this->GetLineNumber(pos,!checkComments);
+      long int linenum = this->GetLineNumber(pos,!checkComments);
       if(linenum != currentline)
         {
         currentline = linenum;
